3.1-drf-intro/smart_home/requirements.txt
django==4.1.1
djangorestframework==3.14.0
psycopg2-binary==2.9.3
python-dotenv==0.21.0
Pillow==9.2.0
